pyproject.toml
[project]
name = "liquidation-bot"
version = "0.1.1"
description = ""
authors = [{ name = "8ball030", email = "8baller@station.codes" }]
readme = "README.md"
requires-python = ">=3.11,<4"
dependencies = [
  "requests>=2.32.5,<3.0.0",
  "eth-utils>=5.3.1,<6.0.0",
  "flask>=3.1.2,<4.0.0",
  "flask-cors>=6.0.1,<7.0.0",
  "web3>=7.13.0,<8.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "python-dotenv>=1.1.1,<2.0.0",
  "apprise>=1.9.4,<2.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.4.2",
  "ruff>=0.13.1",
  "semver>=3.0.4",
  "tbump>=6.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 120
target-version = "py313"
respect-gitignore = true
extend-exclude = [
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".tox",
  ".venv",
  "_build",
  "buck-out",
  "build",
  "dist",
  "docs/conf.py",
  ".nix-venv",
]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
#  "B",   # flake8-bugbear
#  "UP",  # pyupgrade
#  "C4",  # flake8-comprehensions
#  "SIM", # flake8-simplify
#  "PT",  # flake8-pytest-style (if you use pytest)
#  "RUF", # Ruff-specific rules
]
ignore = ["E501"]

[tool.ruff.lint.isort]
split-on-trailing-comma = true

[tool.ruff.format]
# Replace Black
line-ending = "auto"
quote-style = "preserve"
indent-style = "space"
skip-magic-trailing-comma = false
